Stable Commit, can be reintegrated to develop branch from here - Changes made by now: camelCase and Functions out of parser.y
