+++ src/symboltable/funcSymboltable.h
@@ -1,3 +1,2 @@
 #pragma once
 
-void addFunc(char* id, char* type, int numberOfParams);
